Fix : GameListVO, ProfileVO, UserVO type 수정

+++ omok/src/main/java/VO/GameListVO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class GameListVO {
-    private Long gameId;
+    private int gameId;
     private Boolean isCustom;
     private String gameCode;
     private String userId;
